requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
django-cors-headers>=4.3.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
